Ajout d'une colonne ppUser pour ppSpotify + Ajout de l'img dans profil edit et commentaire + suppression commentaire without img

+++ controllers/forum.controller.js
@@ -5,42 +5,38 @@
 // Import Module
 // const { MODE } = process.env
 require("dotenv").config();
-const path = require("path")
-const fs= require ("fs")
-
-
+const path = require("path");
+const fs = require("fs");
 
 //PING POUR TEST UNITAIRE
 exports.getPing = (req, res) => {
   const data = {
-    name: "pong !"
-  }
-  console.log('Controller ping')
+    name: "pong !",
+  };
+  console.log("Controller ping");
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   } else {
-    res.render("home", { data })
+    res.render("home", { data });
   }
-}
-
+};
 
 // Page Forum
 exports.getPageForum = async (req, res) => {
   let data;
 
-  data = await db.query(`SELECT * FROM user INNER JOIN comments ON user.id = comments.id_user`)
-  // console.log('infoData' , data);
-  // TEST UNITAIRE OU VRAI CODE
+  data = await db.query(
+    `SELECT * FROM user INNER JOIN comments ON user.id = comments.id_user`
+  );
+  // TEST UNITAIRE OU PRODUCTION
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   } else {
     res.render("forum", {
-      db: data
-    })
+      db: data,
+    });
   }
-}
-
-
+};
 
 // Envoi du commentaire
 exports.sendComment = async (req, res) => {
@@ -50,30 +46,34 @@ exports.sendComment = async (req, res) => {
   let data;
 
   if (image)
-    data = await db.query(`INSERT INTO comments SET commentary="${commentary}", id_user="${req.session.user.id}" , image="${image}"`)
+    data = await db.query(
+      `INSERT INTO comments SET commentary="${commentary}", id_user="${req.session.user.id}" , image="${image}"`
+    );
   else
-    data = await db.query(`INSERT INTO comments SET commentary="${commentary}", id_user="${req.session.user.id}" , image=''`)
-
+    data = await db.query(
+      `INSERT INTO comments SET commentary="${commentary}", id_user="${req.session.user.id}" , image=''`
+    );
 
   // TEST UNITAIRE OU VRAI CODE
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   } else {
     res.redirect("back");
   }
-}
-
+};
 
 exports.getCommentId = async (req, res) => {
   const { id_comments } = req.params;
   let data;
 
-  data = await db.query(`SELECT * FROM user INNER JOIN comments ON user.id = comments.id_user where id_comments=${id_comments}`)
+  data = await db.query(
+    `SELECT * FROM user INNER JOIN comments ON user.id = comments.id_user where id_comments=${id_comments}`
+  );
   // TEST UNITAIRE OU VRAI CODE
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   }
-}
+};
 
 // Modification du Commentaire
 exports.editComment = async (req, res) => {
@@ -82,34 +82,37 @@ exports.editComment = async (req, res) => {
   let data;
   data = await db.query(`UPDATE comments 
   SET commentary="${newcommentary}" 
-  WHERE id_comments=${id_comments};`)
+  WHERE id_comments=${id_comments};`);
   // TEST UNITAIRE OU VRAI CODE
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   } else {
-    res.redirect("back"), { db: data }
+    res.redirect("back"), { db: data };
   }
-}
-
-
+};
 
 // Suppression du commentaire
 exports.deleteComment = async (req, res) => {
   const { id_comments } = req.params;
- 
-  const [img] = await db.query (`SELECT image FROM comments WHERE id_comments=${id_comments}`)
-  console.log("le nom de mon image", img.image); 
-  if (img.image !== "default.png") {
-    pathImg = path.resolve("public/images/" + img.image)
-    fs.unlink(pathImg, (err) => {
+
+  const [img] = await db.query(
+    `SELECT image FROM comments WHERE id_comments=${id_comments}`
+  );
+  if (img.image !== "") {
+    if (img.image !== "default.png") {
+      pathImg = path.resolve("public/images/" + img.image);
+      fs.unlink(pathImg, (err) => {
         if (err) throw err;
-    })
-}
-  const data = await db.query(`DELETE FROM comments WHERE id_comments=${id_comments}`)
+      });
+    }
+  }
+  const data = await db.query(
+    `DELETE FROM comments WHERE id_comments=${id_comments}`
+  );
 
   if (process.env.MODE === "test") {
-    res.json({ data })
+    res.json({ data });
   } else {
     res.redirect("/forum");
   }
-}
+};
